feat: use exponential backoff on rate limit error
Currently testing out this tap and noticed when running it that I was getting a Runtime error saying it was going over the allowed rate limit. This commit uses what seems to be the common way to handle this exception in singer taps. Specifically this commit adds the backoff page as a dependency of this python package which catches the new RateLimitException exception class. The rate limit exception gets triggered in the method if the request json body includes a specific text string in the message.